utils_peptide/requirements.txt
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astroid==3.0.2
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
biopython==1.81
bleach==6.1.0
Bottleneck==1.3.7
Brotli==1.0.9
cached-property==1.5.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
daal4py==2023.1.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
fonttools==4.45.1
h5py==3.7.0
icecream==2.1.3
idna==3.6
importlib-resources==6.1.1
ipykernel==6.25.0
ipython==8.18.1
ipython-genutils==0.2.0
isort==5.13.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter-server==1.24.0
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
mmcif-pdbx==0+untagged.16.g90d39b2.dirty
modlamp==4.3.0
munkres==1.1.4
mysql-connector-python==8.0.17
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
nose==1.3.7
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.25.2
OpenMM==8.0.0
openpyxl==3.0.10
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pdb2pqr==3.6.1
pdbfixer==1.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3.1
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
ply==3.11
Pmw==2.0.1
ProDy==2.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
propka==3.5.0
protobuf==4.25.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycollada==0.7.2
pycparser==2.21
pydocstyle==6.3.0
Pygments==2.17.2
pykerberos==1.2.4
pylint==3.0.3
pymol==2.5.7
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
rapidfuzz==2.13.7
rdkit==2023.3.3
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
rpds-py==0.13.2
scikit-learn==1.3.0
scikit-learn-intelex==20230426.111436
SciPy==1.11.4
seaborn==0.12.2
Send2Trash==1.8.2
seqeval==1.2.2
setuptools==68.2.2
sip==6.7.12
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
stack-data==0.6.2
tabulate==0.9.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.1.0
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.6.4
wheel==0.42.0
zipp==3.17.0
